+++ src/bot.py
@@ -1,17 +1,18 @@
+import asyncio
 import os
 import sys
-from typing import List
-import discord
-import aiohttp
-import asyncio
 from datetime import datetime, timedelta
+
+import discord
 from discord.ext import commands, tasks
 from dotenv import load_dotenv
 
+from llm import query_ai
+
+sys.path.append(os.path.dirname(__file__))
 # Configure logging to stdout
 import logging
-from shrek_chat import LlmInterface, ShrekChat
-from openrouterllm import OpenRouterLlm
+
 
 logging.basicConfig(
     level=logging.INFO,
@@ -22,7 +23,6 @@ logger = logging.getLogger(__name__)
 
 # Constants
 GENERAL_CHANNEL_ID = 670339732059979807
-AI_ENDPOINT = "http://192.168.0.192:5000/v1/completions"
 MORNING_INTERVAL = 23  # hours
 
 # Load environment variables
@@ -34,28 +34,18 @@ intents = discord.Intents.default()
 intents.message_content = True
 intents.members = True
 
-bot = commands.Bot(command_prefix='!', intents=intents)
 
-async def query_ai(prompt):
-    """Query the AI endpoint with a prompt"""
-    try:
-        payload = {
-            "prompt": prompt,
-            "max_tokens": 200,
-            "temperature": 0.6,
-            "seed": 10
-        }
-        async with aiohttp.ClientSession() as session:
-            async with session.post(AI_ENDPOINT, json=payload) as response:
-                if response.status == 200:
-                    data = await response.json()
-                    return data.get('text', 'No response from AI')
-                else:
-                    logger.error(f"AI API error: {response.status}")
-                    return "Sorry, I couldn't get a response from the AI right now."
-    except Exception as e:
-        logger.error(f"Error querying AI: {e}")
-        return "Sorry, there was an error communicating with the AI."
+class MrBot(commands.Bot):
+    async def setup_hook(self):
+        extensions = [
+            "shrek_chat",
+        ]
+        for extension in extensions:
+            await self.load_extension(extension)
+
+
+bot = MrBot(command_prefix='!', intents=intents)
+
 
 async def check_ai_health():
     """Check if the AI endpoint is responsive"""
@@ -71,6 +61,7 @@ async def check_ai_health():
         logger.error(f"AI endpoint health check failed: {e}")
         return False
 
+
 @tasks.loop(hours=MORNING_INTERVAL)
 async def send_good_morning():
     """Send good morning message every 23 hours"""
@@ -80,6 +71,7 @@ async def send_good_morning():
         logger.info("Sent good morning message")
     else:
         logger.error("Could not find general channel")
+
 
 @send_good_morning.before_loop
 async def before_good_morning():
@@ -92,43 +84,28 @@ async def before_good_morning():
         next_run = next_run + timedelta(days=1)
     await asyncio.sleep((next_run - now).seconds)
 
+
 @bot.event
 async def on_ready():
     logger.info(f'Bot {bot.user} has connected to Discord!')
     logger.info(f'Bot is in {len(bot.guilds)} guilds')
-    
+
     # Check AI endpoint health
     logger.info("Performing AI endpoint health check...")
     is_healthy = await check_ai_health()
-    await load_cogs()
-    
     if is_healthy:
         # Start the good morning task only if AI is healthy
-        send_good_morning.start()
         logger.info("Bot startup complete - AI endpoint is healthy")
     else:
         logger.warning("Bot started but AI endpoint is not responding - some features may not work")
 
-class LocalLlm:
-        async def query_ai(self, prompt: str):
-            return await query_ai(prompt)
 
-
-cogs: List[commands.Cog] = [ShrekChat(bot, OpenRouterLlm())]
-
-
-
-async def load_cogs():
-    for cog in cogs:
-        await bot.add_cog(cog)
-
-
-
+# Shrek Chat Part
 # Run the bot
 if __name__ == '__main__':
     logger.info("Starting bot...")
     token = os.getenv('DISCORD_TOKEN')
-    
+
     # Debug token presence (don't log the actual token!)
     if token:
         logger.info("Discord token found")
@@ -139,4 +116,4 @@ if __name__ == '__main__':
         sys.exit(1)
 
     logger.info("Attempting to start bot with token...")
-    bot.run(token) 
+    bot.run(token)
